# Conflicts:
#	ProsjektOppgaveWebAPI/blogDb.db

+++ ProsjektOppgaveWebAPI/Controllers/PostController.cs
@@ -1,6 +1,7 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using ProsjektOppgaveWebAPI.Models;
-using ProsjektOppgaveWebAPI.Models.ViewModel;
 using ProsjektOppgaveWebAPI.Services;
 
 namespace ProsjektOppgaveWebAPI.Controllers;
@@ -25,12 +26,13 @@ public class PostController : ControllerBase
     
     
     [HttpGet("{id:int}")]
-    public PostViewModel GetPost([FromRoute] int id)
+    public Post? GetPost([FromRoute] int id)
     {
-        return _service.GetPostViewModel(id);
+        return _service.GetPost(id);
     }
     
     
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] Post post)
     {
@@ -40,23 +42,30 @@ public class PostController : ControllerBase
         }
 
         var blog = _service.GetBlog(post.BlogId);
-        if (blog.Status != 0) return BadRequest("This blog is closed for new posts and comments!");
+        if (blog != null && blog.Status != 0) return BadRequest("This blog is closed for new posts and comments!");
         
         await _service.SavePost(post, User);
         return CreatedAtAction("GetPosts", new { id = post.BlogId }, post);
     }
 
     
+    [Authorize]
     [HttpPut("{id:int}")]
     public IActionResult Update([FromRoute] int id, [FromBody] Post post)
     {
         if (id != post.PostId)
             return BadRequest();
 
-        var existingPost = _service.GetPostViewModel(id);
+        var existingPost = _service.GetPost(id);
         if (existingPost is null)
             return NotFound();
-
+        
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (existingPost.OwnerId != userId)
+        {
+            return Unauthorized();
+        }
+        
         _service.SavePost(post, User);
 
         return NoContent();
@@ -66,9 +75,15 @@ public class PostController : ControllerBase
     [HttpDelete("{id:int}")]
     public IActionResult Delete([FromRoute] int id)
     {
-        var post = _service.GetPostViewModel(id);
+        var post = _service.GetPost(id);
         if (post is null)
             return NotFound();
+        
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        if (post.OwnerId != userId)
+        {
+            return Unauthorized();
+        }
 
         _service.DeletePost(id, User);
